bridged

+++ exam_s9/ex2.cu
@@ -29,7 +29,7 @@ thrust::host_vector<int> copy_positive_gpu(const thrust::host_vector<int>& hx)
 {
      const thrust::device_vector<int> dx = hx; // host to device
 
-    const int N = hx.size();
+    const long N = hx.size();
     thrust::device_vector<int> is_pos(N);
     thrust::device_vector<int> map(N);
